# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  user.IsUniqueUsernameResponse:
    properties:
      isUnique:
        example: true
        type: boolean
    type: object
  user.UpdateUserRequestBody:
    properties:
      description:
        example: 野球が好きです！
        type: string
      name:
        example: tanaka taro
        type: string
      profileImage:
        example: https://ik.imagekit.io/your_imagekit_id/image.jpg
        type: string
      username:
        example: user123
        maxLength: 255
        minLength: 1
        type: string
    required:
    - description
    - name
    - username
    type: object
  user.UserResponse:
    properties:
      description:
        example: 野球が好きです！
        type: string
      email:
        example: tanaka@example.com
        type: string
      fileId:
        example: "1234567890"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: tanaka taro
        type: string
      profileImage:
        example: https://ik.imagekit.io/your_imagekit_id/image.jpg
        type: string
      username:
        example: user123
        type: string
    type: object
...
  utils.ApiResponse-user_IsUniqueUsernameResponse:
    properties:
      data:
        $ref: '#/definitions/user.IsUniqueUsernameResponse'
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
  utils.ApiResponse-user_UserResponse:
    properties:
      data:
        $ref: '#/definitions/user.UserResponse'
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
  utils.ErrorBasicResponse:
    properties:
      message:
        example: エラーメッセージ
        type: string
      success:
        example: false
        type: boolean
    type: object
...
info:
  contact: {}
  description: このapiは、ビジターゴーのAPIで、ユーザー、スタジアム、遠征記録、などについて扱います。
  title: ビジターゴーAPI
  version: "1.0"
paths:
...
  /api/admin/user/isUnique/{username}:
    get:
      description: リクエストと同じuserNameが登録済みかチェックする
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: 一意かのフラグ
          schema:
            $ref: '#/definitions/utils.ApiResponse-user_IsUniqueUsernameResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: ユーザーネームの重複チェック
      tags:
      - user
  /api/admin/user/logined:
    get:
      description: ヘッダーのトークンからユーザーを取得する
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.ApiResponse-user_UserResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: ログイン済みの場合、ログインユーザーの情報を取得
      tags:
      - user
  /api/admin/user/update/{userId}:
    put:
      description: userIdが同じユーザーの情報を変更する
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: integer
      - description: userId
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequestBody'
      responses:
        "200":
          description: ユーザー情報
          schema:
            $ref: '#/definitions/utils.ApiResponse-user_UserResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: ユーザー情報変更
      tags:
      - user
  /api/admin/user/userId/{userId}:
    get:
      description: userIdからユーザーを1人取得
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: ユーザー情報
          schema:
            $ref: '#/definitions/utils.ApiResponse-user_UserResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: userIdからユーザー情報取得
      tags:
      - user
  /api/admin/user/username/{username}:
    get:
      description: usernameからユーザーを1人取得
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: ユーザー情報
          schema:
            $ref: '#/definitions/utils.ApiResponse-user_UserResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: usernameからユーザー情報取得
      tags:
      - user
produces:
- application/json
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"